+++ src/UserInterface/PythonCharacterManagerModule.cpp
@@ -98,8 +98,8 @@ PyObject * chrmgrGetVIDInfo(PyObject* poSelf, PyObject* poArgs)
 		TPixelPosition kPPosInst;
 		pkInstBase->NEW_GetPixelPosition(&kPPosInst);
 
-		LONG xInst=kPPosInst.x;
-		LONG yInst=kPPosInst.y;
+		int32_t xInst=kPPosInst.x;
+		int32_t yInst=kPPosInst.y;
 
 		CPythonBackground& rkBG=CPythonBackground::Instance();
 		rkBG.LocalPositionToGlobalPosition(xInst, yInst);
